requirements.txt
streamlit
numpy
pandas
python-dotenv
google-generativeai
pymysql
